user_service/requirements.txt
# FastAPI 核心
fastapi[standard]==0.115.0
uvicorn==0.30.6

# 数据库相关
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1

# 安全与鉴权
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0

# 异步请求（用于服务间通信）
httpx==0.27.2